requirements1.txt
asgiref==3.6.0
async-timeout==4.0.2
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
distlib==0.3.6
Django==4.1.5
django-model-utils==4.3.1
django-query-profiler==0.9
django-rest-framework==0.1.0
django-silk==5.0.2
django-sqlprint-middleware==0.1.4
djangorestframework==3.14.0
filelock==3.9.0
gprof2dot==2022.7.29
gunicorn==20.1.0
idna==3.4
isort==5.11.4
Jinja2==3.1.2
MarkupSafe==2.1.1
mmh3==3.0.0
mo-dots==9.300.22349
mo-future==7.298.22349
mo-imports==7.298.22349
mo-parsing==8.327.22363
mo-sql-parsing==9.328.23003
platformdirs==2.6.2
pycodestyle==2.10.0
Pygments==2.14.0
pygments-pprint-sql==0.1.0
python-dateutil==2.8.2
python-environ==0.4.54
pytz==2022.7.1
redis==4.4.2
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.13
virtualenv==20.17.1
